requirements.txt
dotenv==0.9.9
paho-mqtt==2.1.0
python-dotenv==1.1.0
